requirements.txt
# ===== ScholarGuard Dependencies =====

# ----- LangChain Core -----
langchain>=0.3.0
langgraph>=0.2.0
langchain-community>=0.3.0
langchain-chroma>=0.2.0

# ----- LLM Provider: Groq -----
langchain-groq>=0.2.0
groq>=0.11.0

# ----- Embeddings (local HuggingFace) -----
langchain-huggingface>=0.1.0
sentence-transformers>=3.0.0

# ----- Vector Database -----
chromadb>=0.5.0

# ----- Search & Grounding -----
tavily-python>=0.5.0

# ----- Document Processing -----
pypdf>=4.0.0
Pillow>=10.0.0

# ----- UI Framework -----
streamlit>=1.38.0

# ----- Monitoring -----
langsmith>=0.1.0

# ----- Utilities -----
python-dotenv>=1.0.0
